pyproject.toml
[tool.poetry]
name = "agent-eval-harness"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
weave = "*"
pydantic = "^2.8.2"
datasets = "*"
huggingface-hub = "*"

[tool.poetry.group.swebench]
optional = true

[tool.poetry.group.swebench.dependencies]
python = "^3.11"
weave = "0.50.12"
aiofiles = "24.1.0"
aiohttp = "3.9.5"
aioprocessing = "2.0.1"
aiosignal = "1.3.1"
analytics-python = "1.2.9"
annotated-types = "0.7.0"
anyio = "4.4.0"
attrs = "23.2.0"
backoff = "2.2.1"
beautifulsoup4 = "4.12.3"
certifi = "2024.7.4"
cfgv = "3.4.0"
chardet = "5.2.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
datasets = "2.20.0"
dill = "0.3.8"
distlib = "0.3.8"
distro = "1.9.0"
docker = "7.1.0"
docker-pycreds = "0.4.0"
emoji = "2.12.1"
fastcore = "1.5.54"
filelock = "3.15.4"
frozenlist = "1.4.1"
fsspec = "2024.5.0"
ghapi = "1.0.5"
gitdb = "4.0.11"
gitpython = "3.1.43"
gql = "3.5.0"
graphql-core = "3.2.3"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
huggingface-hub = "0.24.0"
identify = "2.6.0"
idna = "3.7"
janus = "1.0.0"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
mdurl = "0.1.2"
multidict = "6.0.5"
multiprocess = "0.70.16"
nodeenv = "1.9.1"
numpy = "1.26.4"
openai = "1.37.1"
packaging = "24.1"
pandas = "2.2.2"
platformdirs = "4.2.2"
pre-commit = "3.7.1"
protobuf = "5.27.2"
psutil = "6.0.0"
pyarrow = "16.1.0"
pyarrow-hotfix = "0.6"
pydantic = "2.8.2"
pydantic-core = "2.20.1"
pygments = "2.18.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-json-logger = "2.0.7"
pytz = "2024.1"
pyyaml = "6.0.1"
regex = "2024.7.24"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
rich = "13.7.1"
sentry-sdk = "2.11.0"
setproctitle = "1.3.3"
six = "1.16.0"
smmap = "5.0.1"
sniffio = "1.3.1"
soupsieve = "2.5"
tenacity = "9.0.0"
tiktoken = "0.7.0"
tqdm = "4.66.4"
typing-extensions = "4.12.2"
tzdata = "2024.1"
unidiff = "0.7.5"
urllib3 = "2.2.2"
virtualenv = "20.26.3"
wandb = "0.17.5"
werkzeug = "3.0.3"
xxhash = "3.4.1"
yarl = "1.9.4"
swebench = { path = "agent_eval_harness/benchmarks/SWE-bench" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
